requirements.txt
# Core data processing
pandas>=2.0.0
openpyxl>=3.1.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# Database
supabase>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Streamlit app
streamlit>=1.30.0
plotly>=5.18.0
